+++ src/ui/widget/image_canvas/image_canvas.py
@@ -13,7 +13,7 @@ from src.common.god.korm_base import KOrmBase
 from src.models.dto.annotation_category import AnnotationCategory
 from src.core.project_info import ProjectInfo
 from src.common.domain.models.kolo_item import KoloItem
-from src.ui.widget.image_canvas.annotation_list import AnnotationList
+from src.ui.widget.annotation_list.annotation_list import AnnotationList
 from src.ui.widget.image_canvas.annotation_view import AnnotationView
